Move ts data files to new directory
