--- pages/service/events/events.ts
@@ -1,45 +1,0 @@
-import {
-  BMOCImages,
-  OvernighterImages,
-  ServiceDayImages,
-  YSDImages,
-} from "./images";
-
-export const Events = [
-  {
-    title: "RPI Service Day",
-    description:
-      "Assist in helping out service organizations with tasks such as cleaning, organizing, and painting. We will have five different service projects, including upkeeping the Sanctuary for Independent Media facilities, can sorting with Mohawk Hudson Humane Society, painting at Northern Rivers, trail maintenance and upkeep at Frear Park, and Card Making on campus!",
-    date: "Planned for: Oct 5th, 2024",
-    timing: "9:00 am to 2:30 pm",
-    location: "Meet at the Union Horseshoe",
-    images: ServiceDayImages,
-  },
-  {
-    title: "Youth Service Day",
-    description:
-      "Every Spring, the APOEZ chapter works with Frear Park to host an Easter Egg Hunt for children in the greater Troy Area! We assist with egg-filling, cleaning the eggs, taking photos at the event, and other coordination.",
-    date: "Easter",
-    timing: "9:00 am to 2:00 pm",
-    location: "Frear Park",
-    images: YSDImages,
-  },
-  {
-    title: "Biggest Meme on Campus",
-    description:
-      "Every Spring, the APOEZ chapter hosts a large-scale fundraising effort for non-profit organizations both locally and nationwide. In this competition, campaigners advocate for their non-profit of choice through postering, tabling, and making some noise. The month of fundraising is ended with debates, the last chance for the RPI community to vote on who they believe is the winner of the biggest meme on campus! All money donated goes to charities.",
-    date: "Every Spring leading up to and during GM Week",
-    timing: "Tabling throughout the week, debates prior to GM debate",
-    location: "RPI Union",
-    images: BMOCImages,
-  },
-  {
-    title: "Overnighter",
-    description:
-      'Every semester, the APOEZ chapter picks a weekend and stays "overnight" at a local camp to assist in facilities maintenance. Common activities include painting, picking up trash alongside trails, drilling, putting together beds, and cleaning.',
-    date: "Each Fall and Spring Semesters",
-    timing: "Typically the Friday to Sunday",
-    location: "YMCA or Scout Camp",
-    images: OvernighterImages,
-  },
-];
